STMP-560 ProductGroup add Product group view modification

+++ grails-app/domain/uk/co/wonderlane/wlpos/ProductGroup.groovy
@@ -5,7 +5,7 @@ import org.joda.time.LocalDate
 class ProductGroup {
 
     int id
-    Byte retailerId
+    int retailerId
     String name
     boolean hidden
     LocalDate startDate
